\name{DBase}
\alias{DBase}
\docType{data}
\title{ A data base of the type used by this package}
\description{
 A data base with frament size information that can be created by the function transDB
}
\usage{data(DBase)}
\format{
  A data frame with 691 observations on the following 5 variables.
  \describe{
    \item{\code{Marker}}{Genetic marker code}
    \item{\code{Sample}}{Sample code}
    \item{\code{Fragment}}{Fragment sizes scored}
    \item{\code{Date}}{An entry field for the date}
    \item{\code{Plate}}{An entry field for fragment analysis plate}
  }
}
\examples{
data(DBase)
}
\keyword{datasets}
