\name{DBZn}
\alias{DBZn}
\docType{data}
\title{ A data base of the type used by this package with Zostera noltii data}
\description{
 A data base with frament size information that can be created by the function transDB
}
\usage{data(DBZn)}
\format{
  A data frame with 3458 observations on the following 5 variables.
  \describe{
    \item{\code{Marker}}{Genetic marker code}
    \item{\code{Sample}}{Sample code}
    \item{\code{Fragment}}{Fragment sizes scored}
    \item{\code{Date}}{An entry field for the date}
    \item{\code{Plate}}{An entry field for fragment analysis plate} }
}
\author{ Filipe Alberto }
\references{
Alberto F. MsatAllele_1.0: an R package to visualize the binning of microsatellite alleles Journal of Heredity. 100(3):394,397
}
\examples{
data(DBZn)
}
\keyword{datasets}
