\name{getAll}
\alias{getAll}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
internal c++ function
}
\description{
internal c++ function
}
\usage{
getAll(fragRef, frag)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fragRef}{All fragments for a locus}
  \item{frag}{The fragment to be assigned to an allele bin}
}

\value{
An integer relating to the bin assignment of frag
}
\author{
Kevin Keenan
}
\keyword{ internal }
